# admin-open-api/src/main/resources/openapi/components/api/produits.yml
openapi: "3.0.0"

info:
  title: API Produit
  version: "1.0"
  description: Produit API

servers:
  - url: http://localhost:8080/backend-admin/api

tags:
  - name: Produit
    description: Produit specific data.

paths:
  /produits:
    get:
      tags:
        - Produit
      summary: Get All Products
      operationId: getAllProducts
      parameters:
        - name: currentPage
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: sizePage
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: SUCCESS
          content:
            'application/json':
              schema:
                $ref: '../model/produits/ProduitsResultListDTO.yml'
        '404':
          description: NOT FOUND
        '400':
          description: BAD REQUEST
    post:  # post request
      tags:
        - Produit
      summary: Create a new product
      operationId: createProduit
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              $ref: '../model/produits/ProduitDTO.yml'
      responses:
        '201':
          description: CREATED
          content:
            'application/json':
              schema:
                $ref: '../model/produits/ProduitDTO.yml'
        '400':
          description: BAD REQUEST

  /produits-reference:
    get:
      tags:
        - Produit
      summary: Retrieve an product
      operationId: getProduit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: SUCCESS
          content:
            'application/json':
              schema:
                $ref: '../model/produits/ProduitDTO.yml'
        '404':
          description: NOT FOUND
        '400':
          description: BAD REQUEST
    put:
      tags:
        - Produit
      summary: Update a product
      operationId: updateProduit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '../model/produits/ProduitDTO.yml'
      responses:
        '200':
          description: SUCCESS
          content:
            application/json:
              schema:
                $ref: '../model/produits/ProduitDTO.yml'
        '404':
          description: NOT FOUND
        '400':
          description: BAD REQUEST
  /produitsByUser:
    get:
      tags:
        - Produit
      summary: Get All Products by user
      operationId: getProductsByUser
      parameters:
        - name: currentPage
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: sizePage
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: SUCCESS
          content:
            'application/json':
              schema:
                $ref: '../model/produits/ProduitResultListByUserDTO.yml'
        '404':
          description: NOT FOUND
        '400':
          description: BAD REQUEST